QI.9v5kg2b: More refactoring
